[Clang] Move GH202317 test into `warn-implicit-unicode-conversions.cpp`
